// src/actions/domain/enums.ts
/**
 * Auto generated. DO NOT edit manually.
 * Last updated on: Fri, 14 Jul 2023 17:58:27 GMT
 */

/**
 * @enum
 * @readonly
 * @description StkvtWorlds Action Names
 */
export enum StkvtWorldsActionName {
    Balanceobsv = 'balanceobsv',
    Stakeobsv = 'stakeobsv',
    Updateconfig = 'updateconfig',
};